% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_APEX.R
\name{aggregate_APEX}
\alias{aggregate_APEX}
\title{aggregate_APEX}
\usage{
aggregate_APEX(result_APEX, group_by_vars)
}
\arguments{
\item{result_APEX}{Apex result}

\item{group_by_vars}{variable need to be aggregated}
}
\value{
aggregated APEX result
}
\description{
aggregate APEX result by a list of variables
}
