% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_APEX.R
\name{Blockcontrol_APEX}
\alias{Blockcontrol_APEX}
\title{Blockcontrol_APEX}
\usage{
Blockcontrol_APEX(APEXData, case_sensitive = FALSE,
  print_summary = TRUE)
}
\arguments{
\item{APEXData}{filename to be read}

\item{case_sensitive}{variable for case sensitive}

\item{print_summary}{whether print summary}
}
\description{
create a vector which contains all the variable names and variable class appearing in APEX dictionary
}
