% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_APEX.R
\name{colclasses_APEX}
\alias{colclasses_APEX}
\title{colclasses_APEX}
\usage{
colclasses_APEX()
}
\description{
create a vector which contains all the variable names and variable class appearing in APEX dictionary
}
