% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amsr.R
\name{dc.amsr}
\alias{dc.amsr}
\title{Download and Cache an AMSR File}
\usage{
dc.amsr(
  year,
  month,
  day,
  server = "http://data.remss.com/amsr2/bmaps",
  version = "v08",
  destdir = ".",
  destfile,
  force = FALSE,
  dryrun = FALSE,
  debug = getOption("dcDebug", 0)
)
}
\arguments{
\item{year, month, day}{numerical values of the year, month, and day
of the desired dataset. Note that one file is archived per day,
so these three values uniquely identify a dataset.
If \code{day} and \code{month} are not provided but \code{day} is,
then the time is provided in a relative sense, based on the present
date, with \code{day} indicating the number of days in the past.
Owing to issues with timezones and the time when the data
are uploaded to the server, \code{day=3} may yield the
most recent available data. For this reason, there is a
third option, which is to leave \code{day} unspecified, which
works as though \code{day=3} had been given.}

\item{server}{character value giving the base name of the server, used in the
construction of URL queries. Since servers tend to change over time, this
is a good argument to check, when debugging code that once worked but
now fails.}

\item{version}{character string indicating the 'version' name in
the files.}

\item{destdir}{Optional string indicating the directory in which to store
downloaded files. If not supplied, \code{"."} is used, i.e. the data file
is stored in the present working directory.  It can be
helpful to create a top-level directory called \code{data},
with a subdirectory for each file type used (e.g. \code{~/data/met}
to hold meteorological files).}

\item{destfile}{Optional string indicating the name of the file.
If not supplied, then the file name is constructed from the other
parameters of the function call, so that subsequent calls with
the same parameters will yield the same result; this is useful
for caching.}

\item{force}{A logical value that indicates whether to force the download,
even if the pathname constructed from \code{destdir} and \code{destfile}
already exists.}

\item{dryrun}{A logical value that indicates whether to return the constructed
web query, without attempting to download the file. This can be helpful in
designing responses to changing URLs.}

\item{debug}{an integer specifying whether debugging information is
to be printed during processing. The printing is done by
a call to \code{\link{dcDebug}()}.  Setting \code{debug=0}
turns off this form of debugging, while higher values yield
more information. If one \code{dc} function calls another, it
passes the value of \code{debug} but decreased by 1, which means
that the value of \code{debug} controls not just the breadth
of debugging, but also the depth.}
}
\value{
String indicating the full pathname to the downloaded file.
}
\description{
Download and Cache an AMSR File
}
\examples{
\dontrun{
## The download takes several seconds.
f <- dc.amsr(2017, 1, 14) # Jan 14, 2017
library(oce)
d <- read.amsr(f)
plot(d)
mtext(d[["filename"]], side=3, line=0, adj=0)
}
}
\references{
\url{http://images.remss.com/amsr/amsr2_data_daily.html}
provides daily images going back to 2012. Three-day,
monthly, and monthly composites are also provided on that site.
}
\concept{functions related to amsr data}
