% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo.R
\name{dc.argo}
\alias{dc.argo}
\title{Download and Cache an Argo Dataset}
\usage{
dc.argo(longitude, latitude, time, server = "www.usgodae.org",
  destdir = ".", destfile, force = FALSE, dryrun = FALSE,
  debug = getOption("dcDebug", 0))
}
\arguments{
\item{longitude}{Two-element numerical vector holding the limits
of longitude (degrees East) to search for Argo profiles.}

\item{latitude}{Two-element numerical vector holding the
latitude limits (degrees North).}

\item{time}{Two-element character vector holding the
time limits in YYYY-MM-DD format.}

\item{server}{String indicating the server. The default is
\code{"cchdo.ucsd.edu"}, and it is unlikely that other values
will work correctly, unless they are mirrors of this site, using
the same URL structure.}

\item{destdir}{Optional string indicating the directory in which to store
downloaded files. If not supplied, \code{"."} is used, i.e. the data file
is stored in the present working directory.  It can be 
helpful to create a top-level directory called \code{data},
with a subdirectory for each file type used (e.g. \code{~/data/met}
to hold meteorological files).}

\item{destfile}{Optional string indicating the name of the file.
If not supplied, then the file name is constructed from the other
parameters of the function call, so that subsequent calls with
the same parameters will yield the same result; this is useful
for caching.}

\item{force}{A logical value that indicates whether to force the download,
even if the pathname constructed from \code{destdir} and \code{destfile}
already exists.}

\item{dryrun}{A logical value that indicates whether to return the constructed
web query, without attempting to download the file. This can be helpful in
designing responses to changing URLs.}

\item{debug}{an integer specifying whether debugging information is
to be printed during processing. The printing is done by
a call to \code{\link{dcDebug}}.  Setting \code{debug=0}
turns off this form of debugging, while higher values yield
more information. If one \code{dc} function calls another, it
passes the value of \code{debug} but decreased by 1, which means
that the value of \code{debug} controls not just the breadth
of debugging, but also the depth.}
}
\value{
String indicating the full pathname to the downloaded file.
}
\description{


The data are downloaded, and a simple filename convention permits a caching scheme
in which subsequent calls provide access to these downloaded data.
}
\details{
\code{dc.argo} downloads Argo profiles from a USGODAE server [1], by
constructing queries that are devised based on inspection of queries
constructed from the GUI-style webpage [2]. Profiles are specified
in a location-time window.  See \dQuote{Caution} for notes on the
brittleness of the procedure used here.
}
\section{Caution}{

USGODAE does not provide an API for downloading data, and so the
queries used by \code{dc.argo} will not work properly if USGODAE
changes their system. For an example, USGODAE presently looks for
the substring \code{".submit=++Go++"} in the query, but if this
gets switched to \code{".submit=Go"}, then \code{dc.argo} will
fail.  It would certainly be preferable to use
a more formal application-programmer-interface, but until USGODAE
provide such an API, \code{dc.argo} provides a stopgap measure
that avoids the requirement to interact with a web interface using
the mouse and keyboard.
}

\examples{
library(dc)
\dontrun{
library(oce)
## 1. Floats near Nova Scotia on Nov 26, 2017.
n <- dc.argo(lon=c(-70,-30), lat=c(40,60),time=c("2017-11-26","2017-11-26"))
a <- read.argo(n[1])
plotTS(a)
label <- paste(a[["longitude"]][1], "E ",
               a[["latitude"]][1], "N ",
               format(a[["time"]][1], "\%b \%d, \%Y"), sep="")
mtext(label, side=3)
}

}
\references{
1. \url{https://www.usgodae.org}

2. \url{http://www.usgodae.org/cgi-bin/argo_select.pl}
}
\author{
Dan Kelley (2017-11-26)
}
