% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo.R
\name{dc.argoById}
\alias{dc.argoById}
\title{Download and cache an argo file specified by id}
\usage{
dc.argoById(
  id = NULL,
  server,
  destdir = ".",
  destfile,
  ntrials = 3,
  force = FALSE,
  dryrun = FALSE,
  quiet = FALSE,
  debug = getOption("dcDebug", 0)
)
}
\arguments{
\item{id}{character value giving the Argo float id. In this version
of \code{\link[=dc.argoById]{dc.argoById()}}, the value of \code{id} must be a full URL for the netcdf
file, and an error is reported if \code{id} does not start with \verb{ftp://}.}

\item{server}{ignored}

\item{destdir}{character value indicating the directory in which to store
downloaded files. The default value of \code{"."} means to store the
downloaded file in the present working directory.  Set \code{destdir=NULL}
if \code{destfile} is a filename with full path information.
File clutter is reduced by creating a top-level directory called
\code{data}, with subdirectories for various file types; see
\dQuote{Examples}.}

\item{destfile}{optional character value that specifies the name to be used
for the downloaded file. If this is not specified, then a name is determined
from the value of \code{id}.}

\item{ntrials}{integer controlling the number of times to attempt
to download the file. The default value, \code{ntrials=3}, indicates
to try a second time if the first one fails, and a third time if
the second one fails, but to stop with an error, if all three fail.}

\item{force}{A logical value that indicates whether to force the download,
even if the pathname constructed from \code{destdir} and \code{destfile}
already exists.}

\item{dryrun}{A logical value that indicates whether to return the constructed
web query, without attempting to download the file. This can be helpful in
designing responses to changing URLs.}

\item{quiet}{logical value, passed to \code{\link[curl:curl_download]{curl::curl_download()}},
that indicates whether to download silently.  Since some downloads
can be slow, the default value is \code{quiet=FALSE}.}

\item{debug}{an integer specifying whether debugging information is
to be printed during processing. The printing is done by
a call to \code{\link[=dcDebug]{dcDebug()}}.  Setting \code{debug=0}
turns off this form of debugging, while higher values yield
more information. If one \code{dc} function calls another, it
passes the value of \code{debug} but decreased by 1, which means
that the value of \code{debug} controls not just the breadth
of debugging, but also the depth.}
}
\description{
Download and cache an argo file specified by id
}
\examples{
# These examples assume that the ~/data/argo directory exists and is readable.
\dontrun{
library(oce)
# Case 1: id is a URL; server is ignored.
id <- "ftp://ftp.ifremer.fr/ifremer/argo/dac/nmdis/2901633/profiles/R2901633_071.nc"
f <- dc.argoById(id, destdir="~/data/argo")
a <- read.oce(f)
summary(a) # an oce object of class 'argo'
par(mfrow=c(2, 2))
plot(a, which="map")
mtext(a[["time"]], cex=par("cex"))
mtext(gsub(".*/", "", f), cex=par("cex"), line=-1)
plot(a, which="TS")
plot(a, which="temperature profile")
plot(a, which="salinity profile")
}
}
\seealso{
Other functions related to argo data: 
\code{\link{dc.argoIndex}()},
\code{\link{dc.argoSearch}()},
\code{\link{dc.argo}()}
}
\concept{functions related to argo data}
