% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo.R
\name{dc.argoSearch}
\alias{dc.argoSearch}
\title{FIXME: this function is in flux: do not use}
\usage{
dc.argoSearch(
  id = NULL,
  longitude = c(-180, 180),
  latitude = c(-90, 90),
  time = Sys.Date() + c(-7 * 86400, 0),
  server = "ftp://usgodae.org",
  destdir = ".",
  destfile,
  force = FALSE,
  dryrun = FALSE,
  debug = getOption("dcDebug", 0)
)
}
\arguments{
\item{id}{character value holding the ID of a platform or instrument.}

\item{longitude}{numeric vector holding longitude(s).}

\item{latitude}{numeric vector holding latitude(s).}

\item{time}{POSIXct vector holding time(s).}

\item{server}{character value giving the base name of the server, used in the
construction of URL queries. Since servers tend to change over time, this
is a good argument to check, when debugging code that once worked but
now fails.}

\item{destdir}{Optional string indicating the directory in which to store
downloaded files. If not supplied, \code{"."} is used, i.e. the data file
is stored in the present working directory.  It can be
helpful to create a top-level directory called \code{data},
with a subdirectory for each file type used (e.g. \code{~/data/met}
to hold meteorological files).}

\item{destfile}{Optional string indicating the name of the file.
If not supplied, then the file name is constructed from the other
parameters of the function call, so that subsequent calls with
the same parameters will yield the same result; this is useful
for caching.}

\item{force}{A logical value that indicates whether to force the download,
even if the pathname constructed from \code{destdir} and \code{destfile}
already exists.}

\item{dryrun}{A logical value that indicates whether to return the constructed
web query, without attempting to download the file. This can be helpful in
designing responses to changing URLs.}

\item{debug}{an integer specifying whether debugging information is
to be printed during processing. The printing is done by
a call to \code{\link{dcDebug}()}.  Setting \code{debug=0}
turns off this form of debugging, while higher values yield
more information. If one \code{dc} function calls another, it
passes the value of \code{debug} but decreased by 1, which means
that the value of \code{debug} controls not just the breadth
of debugging, but also the depth.}
}
\description{
FIXME: this function is in flux: do not use
}
\seealso{
Other functions related to argo data: 
\code{\link{dc.argoID}()},
\code{\link{dc.argoIndex}()},
\code{\link{dc.argo}()}
}
\concept{functions related to argo data}
