% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coastline.R
\name{dc.coastline}
\alias{dc.coastline}
\title{Download a coastline File}
\usage{
dc.coastline(resolution, item = "coastline", destdir = ".", destfile,
  server = "naturalearth", debug = getOption("oceDebug"))
}
\arguments{
\item{resolution}{A character value specifying the desired resolution. The permitted
choices are \code{"10m"} (for 1:10M resolution, the most detailed),
\code{"50m"} (for 1:50M resolution)
and \code{"110m"} (for 1:110M resolution). If \code{resolution} is not supplied,
\code{"50m"} will be used.}

\item{item}{A character value indicating the quantity to be downloaded.
This is normally one of \code{"coastline"}, \code{"land"}, \code{"ocean"},
\code{"rivers_lakes_centerlines"}, or \code{"lakes"}, but the NaturalEarth
server has other types, and advanced users can discover their names by inspecting
the URLs of links on the NaturalEarth site, and use them for \code{item}.
If \code{item} is not supplied, it defaults to \code{"coastline"}.}

\item{destdir}{Optional string indicating the directory in which to store
downloaded files. If not supplied, \code{"."} is used, i.e. the data file
is stored in the present working directory.  It can be
helpful to create a top-level directory called \code{data},
with a subdirectory for each file type used (e.g. \code{~/data/met}
to hold meteorological files).}

\item{destfile}{Optional string indicating the name of the file.
If not supplied, then the file name is constructed from the other
parameters of the function call, so that subsequent calls with
the same parameters will yield the same result; this is useful
for caching.}

\item{debug}{an integer specifying whether debugging information is
to be printed during processing. The printing is done by
a call to \code{\link{dcDebug}}.  Setting \code{debug=0}
turns off this form of debugging, while higher values yield
more information. If one \code{dc} function calls another, it
passes the value of \code{debug} but decreased by 1, which means
that the value of \code{debug} controls not just the breadth
of debugging, but also the depth.}

\item{force}{A logical value that indicates whether to force the download,
even if the pathname constructed from \code{destdir} and \code{destfile}
already exists.}

\item{dryrun}{A logical value that indicates whether to return the constructed
web query, without attempting to download the file. This can be helpful in
designing responses to changing URLs.}
}
\value{
A character value indicating the filename of the result; if
there is a problem of any kind, the result will be the empty
string.
}
\description{


The data are downloaded, and a simple filename convention permits a caching scheme
in which subsequent calls provide access to these downloaded data.
}
\details{
Constructs a query to the NaturalEarth server [1] to download coastline
data (or lake data, river data, etc) in any of three resolutions.
}
\examples{
\dontrun{
library(dc)
library(oce)
# User must create directory ~/data/coastline first.
# As of September 2016, the downloaded file, named
# "ne_50m_coastline.zip", occupies 443K bytes.
filename <- dc.coastline()
coastline <- read.coastline(filename)
plot(coastline)
}

}
\references{
1. The NaturalEarth server is at \url{http://www.naturalearthdata.com}
}
\seealso{
The work is done with \code{\link[utils]{download.file}}.

The work is done with \code{\link{dc}}.

Other functions that download ocean-related data: \code{\link{dc.amsr}},
  \code{\link{dc.argo}}, \code{\link{dc.hydrography}},
  \code{\link{dc.met}}, \code{\link{dc.topo}},
  \code{\link{dc.woa}}
}
