% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.R
\name{dc.dc}
\alias{dc.dc}
\title{Download and Cache a Dataset}
\usage{
dc.dc(url = NULL, destdir = ".", destfile = NULL, force = FALSE,
  dryrun = FALSE, debug = getOption("dcDebug", 0))
}
\arguments{
\item{destdir}{Optional string indicating the directory in which to store
downloaded files. If not supplied, \code{"."} is used, i.e. the data file
is stored in the present working directory.  It can be 
helpful to create a top-level directory called \code{data},
with a subdirectory for each file type used (e.g. \code{~/data/met}
to hold meteorological files).}

\item{destfile}{Optional string indicating the name of the file.
If not supplied, then the file name is constructed from the other
parameters of the function call, so that subsequent calls with
the same parameters will yield the same result; this is useful
for caching.}

\item{force}{A logical value that indicates whether to force the download,
even if the pathname constructed from \code{destdir} and \code{destfile}
already exists.}

\item{dryrun}{A logical value that indicates whether to return the constructed
web query, without attempting to download the file. This can be helpful in
designing responses to changing URLs.}

\item{debug}{an integer specifying whether debugging information is
to be printed during processing. The printing is done by
a call to \code{\link{dcDebug}}.  Setting \code{debug=0}
turns off this form of debugging, while higher values yield
more information. If one \code{dc} function calls another, it
passes the value of \code{debug} but decreased by 1, which means
that the value of \code{debug} controls not just the breadth
of debugging, but also the depth.}
}
\value{
String indicating the full pathname to the downloaded file.
}
\description{
General function for downloading and caching a dataset. This is called
by e.g. \code{\link{dc.argo}} and the other functions in the \package{dc}
package.

The data are downloaded, and a simple filename convention permits a caching scheme
in which subsequent calls provide access to these downloaded data.
}
