% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g1sst.R
\name{dc.g1sst}
\alias{dc.g1sst}
\title{Download and Cache a G1SST (SST data/model) File}
\usage{
dc.g1sst(year, month, day, lonW, lonE, latS, latN,
  server = "https://coastwatch.pfeg.noaa.gov/erddap/griddap", destdir = ".",
  destfile, force = FALSE, dryrun = FALSE, debug = getOption("dcDebug",
  0))
}
\arguments{
\item{year, month, day}{Numerical values of the year, month, and day
of the desired dataset.}

\item{lonW, lonE}{Numerical values, in range -180 to 180, specifying
the longitudes of the western and eastern boundaries of the focus region.}

\item{latS, latN}{Numerical values, in range -90 to 90, specifying
the latitudes of the southern and northern boundaries of the focus region.}

\item{server}{A string naming the server from which data
are to be acquired.}

\item{destdir}{Optional string indicating the directory in which to store
downloaded files. If not supplied, \code{"."} is used, i.e. the data file
is stored in the present working directory.  It can be
helpful to create a top-level directory called \code{data},
with a subdirectory for each file type used (e.g. \code{~/data/met}
to hold meteorological files).}

\item{destfile}{Optional string indicating the name of the file.
If not supplied, then the file name is constructed from the other
parameters of the function call, so that subsequent calls with
the same parameters will yield the same result; this is useful
for caching.}

\item{force}{A logical value that indicates whether to force the download,
even if the pathname constructed from \code{destdir} and \code{destfile}
already exists.}

\item{dryrun}{A logical value that indicates whether to return the constructed
web query, without attempting to download the file. This can be helpful in
designing responses to changing URLs.}

\item{debug}{an integer specifying whether debugging information is
to be printed during processing. The printing is done by
a call to \code{\link{dcDebug}}.  Setting \code{debug=0}
turns off this form of debugging, while higher values yield
more information. If one \code{dc} function calls another, it
passes the value of \code{debug} but decreased by 1, which means
that the value of \code{debug} controls not just the breadth
of debugging, but also the depth.}
}
\value{
String indicating the full pathname to the downloaded file.
}
\description{


The data are downloaded, and a simple filename convention permits a caching scheme
in which subsequent calls provide access to these downloaded data.
}
\details{
\code{dc.g1sst} downloads data from a G1SST server.
}
\section{History}{

Download URLs working as of 2018-02-07.
}

\examples{
\dontrun{
## Scotian Shelf on Oct 14, 2015.
f <- dc.g1sst(2015, 10, 14, -66, -60, 41, 46) # October 14, 2015
library(oce)
d <- read.g1sst(f)
plot(d, "SST")
mtext(d[["time"]], side=3, line=0, adj=0)
}
}
\references{
\enumerate{
\item source URL-construction tutorial \url{https://coastwatch.pfeg.noaa.gov/erddapinfo/index.html}
\item status page \url{https://coastwatch.pfeg.noaa.gov/erddap/status.html}
}
}
\seealso{
Other functions that download ocean-related data: \code{\link{dc.amsr}},
  \code{\link{dc.argo}}, \code{\link{dc.coastline}},
  \code{\link{dc.hydrography}}, \code{\link{dc.met}},
  \code{\link{dc.topo}}, \code{\link{dc.woa}}
}
