% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.R
\docType{package}
\name{dc}
\alias{dc}
\title{dc: A Package to download and cache data files from the web}
\usage{
dc(
  url = NULL,
  destdir = ".",
  destfile = NULL,
  mode = "wb",
  quiet = FALSE,
  force = FALSE,
  dryrun = FALSE,
  ntrials = 3,
  debug = getOption("dcDebug", 0)
)
}
\arguments{
\item{url}{character value giving the full URL of a file to be downloaded. This is
constructed by specialized functions, e.g. \code{\link[=dc.argo]{dc.argo()}} computes this from
its \code{server} and other arguments, based on reverse engineering the URLs
of Argo files provided by a particular server.}

\item{destdir}{character value indicating the directory in which to store
downloaded files. The default value of \code{"."} means to store the
downloaded file in the present working directory.  Set \code{destdir=NULL}
if \code{destfile} is a filename with full path information.
File clutter is reduced by creating a top-level directory called
\code{data}, with subdirectories for various file types; see
\dQuote{Examples}.}

\item{destfile}{character value indicating the name of the file.
If not supplied, then the file name is constructed from the other
parameters of the function call, so that subsequent calls with
the same parameters will yield the same result; this is useful
for caching.}

\item{mode}{character value passed to \code{\link[curl:curl_download]{curl::curl_download()}},
controlling the mode used in writing the local file. The default
value, \code{mode="wb"}, means to set the write format to binary. This
is necessary on Windows machines, which otherwise will alter newline
characters, ruining file integrity for binary files (such as
netcdf files). The setting is not required on other machines, but
it makes sense to use this default in all instances.  See the
documentation for \code{\link[curl:curl_download]{curl::curl_download()}}.}

\item{quiet}{logical value, passed to \code{\link[curl:curl_download]{curl::curl_download()}},
that indicates whether to download silently.  Since some downloads
can be slow, the default value is \code{quiet=FALSE}.}

\item{force}{A logical value that indicates whether to force the download,
even if the pathname constructed from \code{destdir} and \code{destfile}
already exists.}

\item{dryrun}{A logical value that indicates whether to return the constructed
web query, without attempting to download the file. This can be helpful in
designing responses to changing URLs.}

\item{ntrials}{integer controlling the number of times to attempt
to download the file. The default value, \code{ntrials=3}, indicates
to try a second time if the first one fails, and a third time if
the second one fails, but to stop with an error, if all three fail.}

\item{debug}{an integer specifying whether debugging information is
to be printed during processing. The printing is done by
a call to \code{\link[=dcDebug]{dcDebug()}}.  Setting \code{debug=0}
turns off this form of debugging, while higher values yield
more information. If one \code{dc} function calls another, it
passes the value of \code{debug} but decreased by 1, which means
that the value of \code{debug} controls not just the breadth
of debugging, but also the depth.}
}
\value{
String indicating the full pathname to the downloaded file.
}
\description{
General function for downloading and caching a dataset. This is called
by e.g. \code{\link[=dc.argo]{dc.argo()}} and the other functions in the \code{dc}
package.
}
