% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.R
\name{dcDebug}
\alias{dcDebug}
\title{Possibly print a debugging message}
\usage{
dcDebug(debug = 0, ..., style = "plain", unindent = 0)
}
\arguments{
\item{debug}{an integer, less than or equal to zero for no message, and
greater than zero for increasing levels of debugging.  Values greater than 4
are treated like 4.}

\item{...}{one or several character or other values to be printed, analogous
to the \code{...} argument of \code{cat}.}

\item{style}{either a string or a function. If a string,
it must be \code{"plain"} (the default) for plain text,
\code{"bold"}, \code{"italic"}, \code{"red"}, \code{"green"} or \code{"blue"} (with
obvious meanings).
If \code{style} is a function, it must prepend and postpend the text
with control codes, as in the cyan-coloured example; note that
\CRANpkg{crayon} provides many functions that work well for \code{style}.}

\item{unindent}{Number of levels to un-indent, e.g. it is common to set
this to \code{-1} for messages about entering or exiting a function.}
}
\description{
\code{dcDebug} prints a message, if its first argument exceeds 0.
Many \code{dc} functions decrease the \code{debug} level by 1 when they call other
functions, so the effect is a nesting, with more space for deeper function
level.
}
\examples{
foo <- function(debug=1)
{
   dcDebug(debug, "in foo, about to call bar()\n")
   bar(debug=debug-1)
   dcDebug(debug, "in foo, after calling bar()\n")
}
bar <- function(debug=1)
{
   dcDebug(debug, "in bar()\n")
}
foo(debug=2)
foo(debug=1)
foo(debug=0)

}
\author{
Dan Kelley
}
