% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo.R
\name{decodeTextDate}
\alias{decodeTextDate}
\title{Convert string date to POSIXct time}
\usage{
decodeTextDate(t, tz = "UTC")
}
\arguments{
\item{t}{character value giving the time in yyyymmddHHMMSS format,
as illustrated in \dQuote{Examples}.}

\item{tz}{character string}
}
\value{
POSIXct time
}
\description{
Convert string date to POSIXct time
}
\examples{
decodeTextDate("19970729200300") ## 1997 07 29 20 03 00

}
