% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo.R
\name{download.topo}
\alias{download.topo}
\title{Download and Cache a Topographic File}
\usage{
download.topo(west, east, south, north, resolution, format, server,
  destdir = ".", destfile, force = FALSE, debug = getOption("dacDebug",
  0))
}
\arguments{
\item{west, east}{Longitudes of the western and eastern sides of the box.}

\item{south, north}{Latitudes of the southern and northern sides of the box.}

\item{resolution}{Optional grid spacing, in minutes. If not supplied,
a default value of 4 (corresponding to 7.4km, or 4 nautical
miles) is used. Note that (as of August 2016) the original data are on
a 1-minute grid, which limits the possibilities for \code{resolution}.}

\item{format}{Optional string indicating the type of file to download. If
not supplied, this defaults to \code{"gmt"}. See \dQuote{Details}.}

\item{server}{Optional string indicating the server from which to get the data.
If not supplied, the default
\samp{"https://gis.ngdc.noaa.gov/cgi-bin/public/wcs/etopo1.xyz"}
will be used.}

\item{destdir}{Optional string indicating the directory in which to store
downloaded files. If not supplied, \code{"."} is used, i.e. the data file
is stored in the present working directory.  It can be 
helpful to create a top-level directory called \code{data},
with a subdirectory for each file type used (e.g. \code{~/data/met}
to hold meteorological files).}

\item{destfile}{Optional string indicating the name of the file.
If not supplied, then the file name is constructed from the other
parameters of the function call, so that subsequent calls with
the same parameters will yield the same result; this is useful
for caching.}

\item{force}{A logical value that indicates whether to force the download,
even if the pathname constructed from \code{destdir} and \code{destfile}
already exists.}

\item{debug}{an integer specifying whether debugging information is
to be printed during processing. The printing is done by
a call to \code{\link{dacDebug}}.  Setting \code{debug=0}
turns off this form of debugging, while higher values yield
more information. If one \code{dac} function calls another, it
passes the value of \code{debug} but decreased by 1, which means
that the value of \code{debug} controls not just the breadth
of debugging, but also the depth.}
}
\value{
String indicating the full pathname to the downloaded file.
}
\description{


The data are downloaded, and a simple filename convention permits a caching scheme
in which subsequent calls provide access to these downloaded data.
}
\details{
Data are downloaded (from \samp{https://maps.ngdc.noaa.gov/viewers/wcs-client/}, by
default) and a string containing the full path
to the downloaded file is returned. If \code{destfile} is not supplied,
then the filename is constructed from the query, which means that 
subsequent calls to \code{download.topo}  with identical parameters will
simply return the name of the cached file, assuming the user has not
deleted it in the meantime. 

The data are downloaded with \code{\link[utils]{download.file}}, using a URL
devised from reverse engineering web-based queries constructed by
the default \code{server} used here. Note that the data source is "etopo1",
which is a 1 arc-second file [1,2].

Three values are permitted for \code{format},
each named after the
targets of menu items on the
NOAA website (as of August 2016): (1) \code{"aaigrid"} (for
the menu item "ArcGIS ASCII Grid"), which
yields a text file, (2) \code{"netcdf"} (the default,
for the menu item named
"NetCDF"), which yields a NetCDF file
and (3) \code{"gmt"} (for the menu item named
"GMT NetCDF"), which yields a NetCDF file in
another format. All of these file formats are
recognized by \code{\link[oce]{read.topo}} in the \code{oce} package.
(The NOAA server has more options, and if
\code{\link[oce]{read.topo}} is extended to handle them, they will
also be added here.)
}
\section{Webserver history}{

All versions of \code{download.topo} to date have used a NOAA server as
the data source, but the URL has not been static. A list of the
servers that have been used is provided below,
in hopes that it can help users to make guesses
for \code{server}, should \code{download.topo} fail because of 
a fail to download the data because of a broken link. Another
hint is to look at the source code for
\code{\link[marmap]{getNOAA.bathy}} in the \CRANpkg{marmap} package,
which is also forced to track the moving target that is NOAA.

\itemize{
\item August 2016. 
\samp{http://maps.ngdc.noaa.gov/mapviewer-support/wcs-proxy/wcs.groovy}

\item December 2016.
\samp{http://mapserver.ngdc.noaa.gov/cgi-bin/public/wcs/etopo1.xyz}

\item June-September 2017.
\samp{https://gis.ngdc.noaa.gov/cgi-bin/public/wcs/etopo1.xyz}
}
}

\examples{
\dontrun{
topoFile <- dac::download.topo(west=-64, east=-60, south=43, north=46)
library(oce)
topo <- read.topo(topoFile)
imagep(topo, zlim=c(-400, 400), drawTriangles=TRUE)
data(coastlineWorldFine, package="ocedata")
lines(coastlineWorldFine[["longitude"]], coastlineWorldFine[["latitude"]])
}

}
\references{
1. \samp{https://www.ngdc.noaa.gov/mgg/global/global.html}

2. Amante, C. and B.W. Eakins, 2009. ETOPO1 1 Arc-Minute Global Relief
Model: Procedures, Data Sources and Analysis. NOAA Technical Memorandum
NESDIS NGDC-24. National Geophysical Data Center, NOAA. doi:10.7289/V5C8276M
[access date: Aug 30, 2017].
}
\seealso{
The work is done with \code{\link[utils]{download.file}}.
}
\author{
Dan Kelley
}
