% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_APEX.R
\name{levelCase_APEX}
\alias{levelCase_APEX}
\title{levelCase_APEX}
\usage{
levelCase_APEX(Data)
}
\arguments{
\item{Data}{Apex format data}
}
\description{
convert letter case to the standard APEX format
}
