% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_APEX.R
\name{merge_APEX}
\alias{merge_APEX}
\title{merge_APEX}
\usage{
merge_APEX(Product, Policy)
}
\arguments{
\item{Product}{product dataframe which is loaded by read_APEX}

\item{Policy}{policy dataframe which is loaded by read_APEX}
}
\value{
merged APEX dataframe
}
\description{
merge product and policy file
}
