% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_APEX.R
\name{read_APEX}
\alias{read_APEX}
\title{read_APEX}
\usage{
read_APEX(myfile = NULL, encoding = "UTF-8", case_template = "APEX",
  sheet = 1, guess_max = 50000, Product_file_name = NULL, ...)
}
\arguments{
\item{myfile}{filename to be read}

\item{encoding}{default:"UTF-8" encoding to be assumed for input strings: "unknown", "latin1", "UTF-8" - important for German "Umlaut"!}

\item{case_template}{default: "APEX", or any function like "toupper" or "tolower"}

\item{sheet}{For Excel : Sheet to read in Excel file. Either a string (the name of the sheet) or an integer}

\item{guess_max}{For Excel : Maximum number of rows to use for guessing column types}

\item{Product_file_name}{Product file to be read and merged}

\item{...}{extra named argument passed to read_csv or read_excel}
}
\description{
read csv or excel in APEX format
}
