% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_APEX.R
\name{split_APEX}
\alias{split_APEX}
\title{split_APEX}
\usage{
split_APEX(merge_APEX)
}
\arguments{
\item{merge_APEX}{filename to be read}
}
\description{
create a vector which contains all the variable names and variable class appearing in APEX dictionary
}
