% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_APEX.R
\name{write_APEX}
\alias{write_APEX}
\title{write_APEX}
\usage{
write_APEX(mydata, myfile, fileEncoding = "UTF-8")
}
\arguments{
\item{mydata}{dataframe to be written}

\item{myfile}{saved file name/ path}

\item{fileEncoding}{default: "UTF-8"}
}
\description{
write dataframe to APEX format
}
