% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo.R
\name{dc.argoIndex}
\alias{dc.argoIndex}
\title{Get an index of available floats}
\usage{
dc.argoIndex(
  server = "ftp://usgodae.org/pub/outgoing/argo",
  file = "ar_index_global_prof.txt.gz",
  destdir = ".",
  age = 1/24,
  debug = getOption("dcDebug", 0)
)
}
\arguments{
\item{server}{character value indicating the server.}

\item{file}{character value indicating the file on the server, also
used for the downloaded file, which is placed in the `destdir` direcory.}

\item{destdir}{Optional string indicating the directory in which to store
downloaded files. If not supplied, `"."` is used, i.e. the data file
is stored in the present working directory.  It can be
helpful to create a top-level directory called `data`,
with a subdirectory for each file type used (e.g. `~/data/met`
to hold meteorological files).}

\item{age}{caching age, in days. The default, 1/24, means to re-download
if the local file is more than 1 hour old.  The point of this argument
is to save slow downloads, for index files (even compressed ones)
are several tens of megabytes.  Set `age=0` to force
a download.}

\item{debug}{integer value indicating whether to print debugging messages;
this is passed to [dcDebug()], which handles messages.}
}
\value{
a data frame with columns named `url`, `longitude` and `latitude`.
}
\description{
Constructs a file name and tries to download it to a local file with the
name given as the `file` argument. If there is already a local file, and if
it is less than `age` days old, then the local file is used and the download
is skipped. In either case, this function reads that file, skips the header,
and returns a data frame holding `url` (the spot from which the file
can be downloaded), `longitude` (the listed longitude of the float)
and `latitude` (the listed latitude of the float).
}
\examples{
\dontrun{
d <- dc.argoIndex() # ftp://usgodae.org/pub/outgoing/argo/ar_index_global_meta.txt
names(d) # reveals 'longitude' and 'latitude', plus other things

library(oce)
data(coastlineWorld)
par(mar=rep(0.1, 4))
mapPlot(coastlineWorld, projection="+proj=moll", col="lightgray")
mapPoints(d$longitude, d$latitude, pch=".")
}

}
\seealso{
Other functions related to argo data: 
\code{\link{dc.argoID}()},
\code{\link{dc.argoSearch}()},
\code{\link{dc.argo}()}
}
\author{
Dan Kelley
}
\concept{functions related to argo data}
