% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coastline.R
\name{dc.coastline}
\alias{dc.coastline}
\title{Download and Cache a Coastline File}
\usage{
dc.coastline(
  resolution,
  item = "coastline",
  server = "naturalearth",
  destdir = ".",
  destfile,
  force = FALSE,
  dryrun = FALSE,
  debug = getOption("oceDebug")
)
}
\arguments{
\item{resolution}{A character value specifying the desired resolution. The permitted
choices are \code{"10m"} (for 1:10M resolution, the most detailed),
\code{"50m"} (for 1:50M resolution)
and \code{"110m"} (for 1:110M resolution). If \code{resolution} is not supplied,
\code{"50m"} will be used.}

\item{item}{A character value indicating the quantity to be downloaded.
This is normally one of \code{"coastline"}, \code{"land"}, \code{"ocean"},
\code{"rivers_lakes_centerlines"}, or \code{"lakes"}, but the NaturalEarth
server has other types, and advanced users can discover their names by inspecting
the URLs of links on the NaturalEarth site, and use them for \code{item}.
If \code{item} is not supplied, it defaults to \code{"coastline"}.}

\item{server}{character value giving the base name of the server, used in the
construction of URL queries. Since servers tend to change over time, this
is a good argument to check, when debugging code that once worked but
now fails.}

\item{destdir}{character value indicating the directory in which to store
downloaded files. The default value of \code{"."} means to store the
downloaded file in the present working directory.  Set \code{destdir=NULL}
if \code{destfile} is a filename with full path information.
File clutter is reduced by creating a top-level directory called
\code{data}, with subdirectories for various file types; see
\dQuote{Examples}.}

\item{destfile}{character value indicating the name of the file.
If not supplied, then the file name is constructed from the other
parameters of the function call, so that subsequent calls with
the same parameters will yield the same result; this is useful
for caching.}

\item{force}{A logical value that indicates whether to force the download,
even if the pathname constructed from \code{destdir} and \code{destfile}
already exists.}

\item{dryrun}{A logical value that indicates whether to return the constructed
web query, without attempting to download the file. This can be helpful in
designing responses to changing URLs.}

\item{debug}{an integer specifying whether debugging information is
to be printed during processing. The printing is done by
a call to \code{\link[=dcDebug]{dcDebug()}}.  Setting \code{debug=0}
turns off this form of debugging, while higher values yield
more information. If one \code{dc} function calls another, it
passes the value of \code{debug} but decreased by 1, which means
that the value of \code{debug} controls not just the breadth
of debugging, but also the depth.}
}
\value{
A character value indicating the filename of the result; if
there is a problem of any kind, the result will be the empty
string.
}
\description{
Constructs a query to the NaturalEarth server (ref 1) to download coastline
data (or lake data, river data, etc) in any of three resolutions,
and caches the resultant file.
}
\examples{
\dontrun{
library(dc)
library(oce)
# User must create directory ~/data/coastline first.
# As of September 2016, the downloaded file, named
# "ne_50m_coastline.zip", occupies 443K bytes.
filename <- dc.coastline(destdir="~/data/coastline")
coastline <- read.coastline(filename)
plot(coastline)
}

}
\references{
\enumerate{
\item The NaturalEarth server is at \url{http://www.naturalearthdata.com}
}
}
\seealso{
The work is done with \code{\link[utils:download.file]{utils::download.file()}}.

The work is done with \code{\link[=dc]{dc()}}.

Other functions that download ocean-related data: 
\code{\link{dc.g1sst}()},
\code{\link{dc.hydrography}()},
\code{\link{dc.met}()},
\code{\link{dc.topo}()},
\code{\link{dc.woa}()}
}
\concept{functions that download ocean-related data}
