% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrography.R
\name{dc.hydrography}
\alias{dc.hydrography}
\title{Download and Cache a Hydrographic Dataset}
\usage{
dc.hydrography(expocode, type = "bottle", format = "exchange", server,
  destdir = ".", destfile, force = FALSE, dryrun = FALSE,
  debug = getOption("dcDebug", 0))
}
\arguments{
\item{expocode}{String indicating the cruise (or 'expedition') code.}

\item{type}{String indicating data type, either \code{"bottle"} (the
default) or \code{"ctd"}.}

\item{format}{String indicating data format, either \code{"exchange"}
(the default) or \code{"whp_netcdf"}.}

\item{server}{String indicating the server. The default is
\code{"cchdo.ucsd.edu"}, and it is unlikely that other values
will work correctly, unless they are mirrors of this site, using
the same URL structure.}

\item{destdir}{Optional string indicating the directory in which to store
downloaded files. If not supplied, \code{"."} is used, i.e. the data file
is stored in the present working directory.  It can be 
helpful to create a top-level directory called \code{data},
with a subdirectory for each file type used (e.g. \code{~/data/met}
to hold meteorological files).}

\item{destfile}{Optional string indicating the name of the file.
If not supplied, then the file name is constructed from the other
parameters of the function call, so that subsequent calls with
the same parameters will yield the same result; this is useful
for caching.}

\item{force}{A logical value that indicates whether to force the download,
even if the pathname constructed from \code{destdir} and \code{destfile}
already exists.}

\item{dryrun}{A logical value that indicates whether to return the constructed
web query, without attempting to download the file. This can be helpful in
designing responses to changing URLs.}

\item{debug}{an integer specifying whether debugging information is
to be printed during processing. The printing is done by
a call to \code{\link{dcDebug}}.  Setting \code{debug=0}
turns off this form of debugging, while higher values yield
more information. If one \code{dc} function calls another, it
passes the value of \code{debug} but decreased by 1, which means
that the value of \code{debug} controls not just the breadth
of debugging, but also the depth.}
}
\value{
String indicating the full pathname to the downloaded file.
}
\description{


The data are downloaded, and a simple filename convention permits a caching scheme
in which subsequent calls provide access to these downloaded data.
}
\details{
\code{dc.hydography} downloads hydographic datasets from the CCHDO archive,
(or mirrors that use the exact same URL structure, apart from the 
server name) based on a specified "expocode" that specifies the
expedition or cruise. This is convenient, because expocodes are how
cruises are referred to in the literature. See \dQuote{Limitiations} for
a comments on the brittleness of \code{dc.hydrography}.
}
\section{Limitations}{

Unfortunately, \url{cchdo.ucsd.edu} does not provide data organized
into URLs that can be inferred from the expocode alone, and so the
procedure used here is to scan the HTML for a summary URL that
\emph{can} be inferred, and then to discover links from the HTML
text. This is brittle to changes in the formatting used by CCHDO, which
means that changes may need to be made to \code{dc.hydography} whenever
CCHDO changes their website. Another consequence is that a web connection
is required even if the desired file has already been downloaded.
}

\examples{
library(dc)
\dontrun{
## 1. A cruise between Australia and Antarctica.
d <- dc.hydrography("09AR20041223")
library(oce)
plot(read.section(d), which="map")
}

}
\references{
1. \url{https://www.nodc.noaa.gov/OC5/woa13/woa13data.html}

2. \url{https://www.nodc.noaa.gov/OC5/woa13}
}
\seealso{
The work is done with \code{\link[utils]{download.file}}.
}
\author{
Dan Kelley (2017-09-22)
}
