% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrography.R
\name{dc.hydrography}
\alias{dc.hydrography}
\title{Download and Cache a Hydrographic Dataset}
\usage{
dc.hydrography(
  expocode,
  type = "bottle",
  format = "exchange",
  server,
  destdir = ".",
  destfile,
  force = FALSE,
  dryrun = FALSE,
  debug = getOption("dcDebug", 0)
)
}
\arguments{
\item{expocode}{String indicating the cruise (or 'expedition') code.}

\item{type}{String indicating data type, either `"bottle"` (the
default) or `"ctd"`.}

\item{format}{String indicating data format, either `"exchange"`
(the default) or `"whp_netcdf"`.}

\item{server}{character value giving the base name of the server, used in the
construction of URL queries. Since servers tend to change over time, this
is a good argument to check, when debugging code that once worked but
now fails.}

\item{destdir}{Optional string indicating the directory in which to store
downloaded files. If not supplied, `"."` is used, i.e. the data file
is stored in the present working directory.  It can be
helpful to create a top-level directory called `data`,
with a subdirectory for each file type used (e.g. `~/data/met`
to hold meteorological files).}

\item{destfile}{Optional string indicating the name of the file.
If not supplied, then the file name is constructed from the other
parameters of the function call, so that subsequent calls with
the same parameters will yield the same result; this is useful
for caching.}

\item{force}{A logical value that indicates whether to force the download,
even if the pathname constructed from `destdir` and `destfile`
already exists.}

\item{dryrun}{A logical value that indicates whether to return the constructed
web query, without attempting to download the file. This can be helpful in
designing responses to changing URLs.}

\item{debug}{an integer specifying whether debugging information is
to be printed during processing. The printing is done by
a call to [dcDebug()].  Setting `debug=0`
turns off this form of debugging, while higher values yield
more information. If one `dc` function calls another, it
passes the value of `debug` but decreased by 1, which means
that the value of `debug` controls not just the breadth
of debugging, but also the depth.}
}
\value{
String indicating the full pathname to the downloaded file.
}
\description{
[dc.hydography()] downloads hydographic datasets from the CCHDO archive,
(or mirrors that use the exact same URL structure, apart from the
server name) based on a specified "expocode" that specifies the
expedition or cruise. This is convenient, because expocodes are how
cruises are referred to in the literature. See \dQuote{Limitiations} for
a comments on the brittleness of [dc.hydrography)(].
}
\section{Limitations}{

Unfortunately, \url{cchdo.ucsd.edu} does not provide data organized
into URLs that can be inferred from the expocode alone, and so the
procedure used here is to scan the HTML for a summary URL that
\emph{can} be inferred, and then to discover links from the HTML
text. This is brittle to changes in the formatting used by CCHDO, which
means that changes may need to be made to [dc.hydography()] whenever
CCHDO changes their website. Another consequence is that a web connection
is required even if the desired file has already been downloaded.
}

\examples{
library(dc)
\dontrun{
library(oce)
## 1. Bottle stations for a Southern Ocean cruise
bottle <- dc.hydrography("09AR20041223")
plot(read.section(bottle), which="map")
##
## 2. CTD data from same cruise, subsetted to
##    a section between Antarctica and Australia
ctd <- dc.hydrography("09AR20041223", type="ctd")
secCTD <- read.section(directory=ctd)
transectCTD <- subset(secCTD, 100 < longitude & longitude < 120)
plot(transectCTD)
}

}
\references{
1. \url{https://cchdo.ucsd.edu}
}
\seealso{
The work is done with [dc()].

Other functions that download ocean-related data: 
\code{\link{dc.coastline}()},
\code{\link{dc.g1sst}()},
\code{\link{dc.met}()},
\code{\link{dc.topo}()},
\code{\link{dc.woa}()}
}
\author{
Dan Kelley (2017-11-25)
}
\concept{functions that download ocean-related data}
